Adding more jsdoc documentation. Rearranging exports to make them jsdoc friendly. Remove the jsdoc date. Make sure npm run doc removes old residues.
