+++ src/plotOptions.js
@@ -5,17 +5,31 @@ import { ClusterMapOptions } from './algorithm';
 
 let currentPlot = null;
 
+// >>> OPTIONS
+
 /**
- * @summary A map of valid options with option descriptions,
- *          validation functions and flags about side effects
+ * A map of valid options with option descriptions,
+ * validation functions and flags about side effects
  * @const
- * @enum {OptionDescription}
+ * @enum {{
+ * default: *,
+ * valid: (function(*)|Array),
+ * requireRedraw: boolean,
+ * requireRecompile: boolean
+ * }}
  */
-export const Options = {
+const OPTIONS = {
   // General plot options
 
-  /** The space around the drawing area in the form [top, right, bottom, left].
-   * X-axis, y-axis and colormap are drawn within padding space. */
+  /**
+   * The space around the drawing area. X-axis, y-axis and colormap are drawn within padding space.
+   * Values can be specified as array of length 4 for [top, right, bottom, left],
+   * or as a single value specifying all four corners.
+   * Values can be absolute (e.g. 120 or "120" ) or
+   * percentages (e.g. "20%"" or ["10%", "20%", "10%", "20%""]
+   * (default: [50, 60, 50, 50])
+   * @type {number[]|string[]|number|string}
+   */
   padding: {
     description: 'The space around the drawing area in the form [top, right, bottom, left].' +
       ' X-axis, y-axis and colormap are drawn within padding space.',
@@ -26,7 +40,9 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
-  /** When enabled, shows a colormap to the right of the plot. */
+
+  /** When enabled, shows a colormap to the right of the plot.
+   * @type {boolean} */
   showColormap: {
     description: 'When enabled, shows a colormap to the right of the plot.',
     default: true,
@@ -34,7 +50,10 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
-  /** When enabled, scrolling above the plot zooms in or out of the data. */
+
+  /** When enabled, scrolling above the plot zooms in or out of the data. (default: true)
+   * @type {boolean}
+   */
   enableScrolling: {
     description: 'When enabled, scrolling above the plot zooms in or out of the data.',
     default: true,
@@ -42,7 +61,10 @@ export const Options = {
     requireRedraw: false,
     requireRecompile: false,
   },
-  /** When enabled, thumbnails can be dragged with the mouse. */
+
+  /** When enabled, thumbnails can be dragged with the mouse. (default: true)
+   * @type {boolean}
+   */
   enableThumbnailDragging: {
     description: 'When enabled, thumbnails can be dragged with the mouse.',
     default: true,
@@ -52,8 +74,11 @@ export const Options = {
   },
 
   // Advanced plot options
-  /** When enabled, the canvas is continuously rerendered at up to 60 frames per second.
-   *  Keep this setting disabled to save processing resources. */
+  /**
+   * When enabled, the canvas is continuously rerendered at up to 60 frames per second.
+   * Keep this setting disabled to save processing resources. (default: false)
+   * @type {boolean}
+   */
   enableContinuousRendering: {
     description: 'When enabled, the canvas is continuously rerendered at up to 60 frames ' +
       'per second. Keep this setting disabled to save processing resources.',
@@ -62,8 +87,12 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
-  /** Enables/disables blending in WebGL. Whenever using any kind of transparency,
-   * this setting should be kept enabled. */
+
+  /**
+   * Enables/disables blending in WebGL. Whenever using any kind of transparency,
+   * this setting should be kept enabled. (default: true)
+   * @type {boolean}
+   */
   enableTransparency: {
     description: 'Enables/disables blending in WebGL. Whenever using any kind of transparency' +
       ', this setting should be kept enabled.',
@@ -72,8 +101,11 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
+
   /** When enabled, draws an image into the background, that shows density of points.
-   * (can be combined with 'showPointClusters') */
+   * (can be combined with 'showPointClusters') (default: false)
+   * @type {boolean}
+   */
   showPointDensity: {
     description: 'When enabled, draws an image into the background,' +
       ' that shows density of points. (can be combined with \'showPointClusters\')',
@@ -82,8 +114,11 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
+
   /** When enabled, draws an image into the background, that shows colored clusters of points.
-   * (can be combined with 'showPointDensity') */
+   * (can be combined with 'showPointDensity')  (default: false)
+   * @type {boolean}
+   */
   showPointClusters: {
     description: 'When enabled, draws an image into the background, ' +
     'that shows colored clusters of points. (can be combined with \'showPointDensity\')',
@@ -92,6 +127,13 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
+
+  /**
+   * Specifies the density threshold ratio.
+   * Densities below threshold * maximum-density are considered outliers. (default: 0.1).
+   * See {@link ClusterMapOptions}
+   * @type {number}
+   */
   pointClusterThreshold: {
     description: 'Controls the realtive threshold between clusters and outliers when' +
     ' showing clusters (see \'showPointClusters\')',
@@ -102,7 +144,9 @@ export const Options = {
   },
 
   // Histogram options
-  /** When enabled, shows a histogram between the x-axis and the plot. */
+  /** When enabled, shows a histogram between the x-axis and the plot. (default: false)
+   * @type {boolean}
+   */
   showXAxisHistogram: {
     description: 'When enabled, shows a histogram between the x-axis and the plot.',
     default: false,
@@ -110,7 +154,10 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
-  /** When enabled, shows a histogram between the y-axis and the plot. */
+
+  /** When enabled, shows a histogram between the y-axis and the plot. (default: false)
+   * @type {boolean}
+   */
   showYAxisHistogram: {
     description: 'When enabled, shows a histogram between the y-axis and the plot.',
     default: false,
@@ -118,7 +165,10 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
-  /** When enabled, shows a histogram between the colormap and the plot. */
+
+  /** When enabled, shows a histogram between the colormap and the plot. (default: false)
+   * @type {boolean}
+   */
   showColormapHistogram: {
     description: 'When enabled, shows a histogram between the colormap and the plot.',
     default: false,
@@ -134,7 +184,10 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
-  /** Controls the height of each histogram in the scatterplot (in pixels). */
+
+  /** Controls the height of each histogram in the scatterplot (in pixels). (default: 64)
+   * @type {number}
+   */
   histogramHeight: {
     description: 'Controls the height of each histogram in the scatterplot (in pixels).',
     default: 64,
@@ -144,7 +197,12 @@ export const Options = {
   },
 
   // Point options
-  /** Controls the shape of data points in the scatterplot. */
+
+  /** Controls the shape of data points in the scatterplot.
+   * Acceptable values are 'Rectangle', 'Circle', 'Cross', 'Diamond', 'Gaussian', 'Custom'.
+   * (default: 'Circle')
+   * @type {string}
+  */
   pointShape: {
     description: 'Controls the shape of data points in the scatterplot.',
     default: 'Circle',
@@ -152,8 +210,11 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: true,
   },
+
   /** When 'pointShape' is set to 'Custom', this defines a GLSL function given vec2 p,
-   * that returns opacity in the range [0.0 ... 1.0] at location p. */
+   * that returns opacity in the range [0.0 ... 1.0] at location p. (default: '{ return 1.0; }')
+   * @type {string}
+   */
   customPointShape: {
     description: 'When \'pointShape\' is set to \'Custom\', this defines a GLSL function ' +
       'given vec2 p, that returns opacity in the range [0.0 ... 1.0] at location p.',
@@ -162,7 +223,10 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: true,
   },
-  /** Controls the diameter of data points in the scatterplot (in pixels). */
+
+  /** Controls the diameter of data points in the scatterplot (in pixels). (default: 6)
+   * @type {number}
+   */
   pointSize: {
     description: 'Controls the diameter of data points in the scatterplot (in pixels).',
     default: 6,
@@ -170,8 +234,11 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
+
   /** Controls the visibility of data points in the scatterplot between
-   *  0 (invisible) and 1 (fully opaque). */
+   *  0 (invisible) and 1 (fully opaque). (default: 1.0)
+   * @type {number}
+   */
   pointOpacity: {
     description: 'Controls the visibility of data points in the scatterplot between ' +
       '0 (invisible) and 1 (fully opaque).',
@@ -180,8 +247,12 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
+
   /** Controls the color of data points in the scatterplot.
-   * Valid values are an array of bytes in RGBA order or a colormap name. */
+   * Valid values are an array of bytes ( 0 .. 255) in RGBA order, a color name (e.g. "red"),
+   * or a colormap name ("exhue" or "rainbow"). (default: "exhue")
+   * @type {(number[] | string)}
+   */
   pointColor: {
     description: 'Controls the color of data points in the scatterplot. ' +
     'Valid values are an array of bytes in RGBA order or a colormap name.',
@@ -192,7 +263,10 @@ export const Options = {
   },
 
   // Thumbnail options
-  /** Controls the width/height of thumbnails in the scatterplot (in pixels). */
+
+  /** Controls the width/height of thumbnails in the scatterplot (in pixels). (default: 64)
+   * @type {number}
+   */
   thumbnailSize: {
     description: 'Controls the width/height of thumbnails in the scatterplot (in pixels).',
     default: 64,
@@ -200,7 +274,10 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
-  /** Controls the width of thumbnail borders in the scatterplot. */
+
+  /** Controls the width of thumbnail borders in the scatterplot. (default: 1)
+   * @type {number}
+   */
   thumbnailBorderWidth: {
     description: 'Controls the width of thumbnail borders in the scatterplot.',
     default: 1,
@@ -208,9 +285,12 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
+
   /** Controls the color of thumbnail borders in the scatterplot.
    * Valid values are an array of bytes in RGBA order, a color name or 'null'.
-   * If set to 'null', the CSS foreground color will be used. */
+   * If set to 'null', the CSS foreground color will be used. (default: null)
+   * @type {(number[] | string)}
+   */
   thumbnailBorderColor: {
     description: 'Controls the color of thumbnail borders in the scatterplot. ' +
       'Valid values are an array of bytes in RGBA order, a color name or \'null\'. ' +
@@ -222,7 +302,9 @@ export const Options = {
   },
   /** Controls the color of thumbnail line in the scatterplot.
    * Valid values are an array of bytes in RGBA order, a color name or 'null'.
-   * If set to 'null', the CSS foreground color will be used. */
+   * If set to 'null', the CSS foreground color will be used. (default: null)
+   * @type {(number[] | string)}
+   */
   thumbnailLineColor: {
     description: 'Controls the color of thumbnail line in the scatterplot. ' +
       'Valid values are an array of bytes in RGBA order, a color name or\'null\'. ' +
@@ -232,9 +314,12 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
+
   /** Controls the color of thumbnail labels in the scatterplot.
    * Valid values are an array of bytes in RGBA order, a color name or 'null'.
-   * If set to 'null', the CSS background color will be used. */
+   * If set to 'null', the CSS background color will be used. (default: null)
+   * @type {(number[] | string)}
+   */
   thumbnailLabelColor: {
     description: 'Controls the color of thumbnail labels in the scatterplot. ' +
       'Valid values are an array of bytes in RGBA order, a color name or \'null\'. ' +
@@ -244,7 +329,11 @@ export const Options = {
     requireRedraw: true,
     requireRecompile: false,
   },
-  /** When enabled, links thumbnails to points using unique labels instead of lines. */
+
+  /** When enabled, links thumbnails to points using unique labels instead of lines.
+   * (default: false)
+   * @type {boolean}
+   */
   labelThumbnails: {
     description: 'When enabled, links thumbnails to points using unique labels instead of lines.',
     default: false,
@@ -254,6 +343,10 @@ export const Options = {
   },
 };
 
-export function setCurrentPlot(plot) {
+// sets the currently active plot (when multiple GlobalView plots are created.)
+// used internally by GlobalView.
+function setCurrentPlot(plot) {
   currentPlot = plot;
 }
+
+export { OPTIONS, setCurrentPlot };
